ingest-api/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
asyncpg==0.30.0
pgvector==0.3.6

# HTTP client
httpx==0.28.1

# Validation and config
pydantic==2.10.3
pydantic-settings==2.6.1

# URL content extraction
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.4.1
